requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2

# Supabase
supabase==2.7.4

# Google Calendar
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.143.0

# Twilio
twilio==9.3.3

# Date/timezone
pytz==2024.1
python-dateutil==2.9.0

# Security
python-jose[cryptography]==3.3.0

# AI APIs
openai==1.50.0

# Validation
phonenumbers==8.13.35
